requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
idna==2.9
psycopg2==2.8.5
pycryptodomex==3.9.7
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stream-chat==1.1.0
urllib3==1.25.8
